feat(marketing): rename Playbooks to Registry + soften homepage vendor line
- Overwrite /registry with the rich playbooks catalog (37 tiles, stats, composability, cross-agent). Keeps existing /playbooks -> /registry redirect.
- Delete standalone /playbooks page — now unreachable behind the redirect.
- Homepage line 560: swap "Cortex enforces inside Cortex. Copilot Studio inside Copilot. Bedrock inside Bedrock…" for the softer "Native platform controls stay in place. Conduct applies one policy and evidence model across the mix of agent tools your team actually runs."

Fixes the "No automation packs found" empty state and removes the named-incumbent framing on the homepage.
